+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## 1.0.9
+
+---
+
 ## 1.0.8
 
 ### Improvements